Made some Changes in Event Module, Added category method to get category wise events

+++ ticketbooking/src/main/java/com/event/ticketbooking/model/User.java
@@ -9,7 +9,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userId;
+    private Long userId;
 
     private String fullName;
     private String email;
@@ -17,19 +17,16 @@ public class User {
     private String phone;
     private String role;
     private String accountStatus;
-
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
     public User() {}
 
-    // ✅ GETTERS & SETTERS
-
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -96,5 +93,4 @@ public class User {
     public void setUpdatedAt(LocalDateTime updatedAt) {
         this.updatedAt = updatedAt;
     }
-
 }
